backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astroid==2.4.2
attrs==23.2.0
billiard==4.2.1
boto3==1.16.59
botocore==1.19.59
celery==5.5.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.4
confusable-homoglyphs==3.3.1
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20230430
dj-database-url==0.5.0
dj-rest-auth==5.0.2
Django==5.2.1
django-admin-adminlte==1.0.5
django-admin-interface==0.28.6
django-allauth==0.61.1
django-api-generator==1.0.10
django-bootstrap-form==3.4
django-colorfield==0.11.0
django-cors-headers==3.5.0
django-crispy-forms==1.11.2
django-dynamic-datatb==1.0.22
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.1
django-guardian==2.4.0
django-heroku==0.3.1
django-import-export==3.3.7
django-jazzmin==2.6.1
django-model-utils==4.4.0
django-notifications-hq==1.8.3
django-psycopg2-extension==0.2.1
django-registration==3.4
django-request-logging==0.7.5
django-reset-migrations==0.4.0
django-restql==0.15.4
django-s3direct==1.1.8
django-seed==0.3.1
django-storages==1.11.1
django-survey-and-report==1.4.8
django-tastypie==0.14.6
djangorestframework==3.15.2
djangorestframework-guardian==0.3.0
djangorestframework-simplejwt==5.3.1
dotty-dict==1.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
Faker==24.3.0
fonttools==4.53.1
gunicorn==20.0.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
isort==5.7.0
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.5.4
lazy-object-proxy==1.4.3
libgravatar==1.0.4
MarkupPy==1.14
matplotlib==3.9.1
mccabe==0.6.1
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
ordereddict==1.1
packaging==24.0
pandas==2.2.1
phonenumbers==8.13.32
pillow==10.4.0
prompt_toolkit==3.0.51
psycopg2==2.9.3
psycopg2-binary==2.8.6
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.8.0
pylint==2.6.0
pyotp==2.9.0
pyparsing==3.1.2
pyPEG2==2.15.2
python-dateutil==2.9.0.post0
python-dotenv==0.15.0
python-mimeparse==1.6.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
s3transfer==0.3.4
setuptools==68.2.2
six==1.15.0
sniffio==1.3.1
sorl-thumbnail==12.10.0
sqlparse==0.4.1
survey==5.2.4
svix==1.69.0
swagger-spec-validator==3.0.3
swapper==1.3.0
tablib==3.5.0
text-unidecode==1.3
toml==0.10.2
toposort==1.10
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.2
vine==5.1.0
wcwidth==0.2.13
wheel==0.41.2
whitenoise==6.7.0
wrapt==1.17.2
xlrd==2.0.1
xlwt==1.3.0
